+++ app/src/main/java/hexlet/code/service/UserService.java
@@ -2,8 +2,10 @@ package hexlet.code.service;
 
 import hexlet.code.dto.UserCreateDTO;
 import hexlet.code.dto.UserUpdateDTO;
+import hexlet.code.exception.ResourceConflictException;
 import hexlet.code.exception.ResourceNotFoundException;
 import hexlet.code.model.User;
+import hexlet.code.repository.TaskRepository;
 import hexlet.code.repository.UserRepository;
 import java.util.List;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -14,10 +16,17 @@ public class UserService {
 
     private final UserRepository userRepository;
 
+    private final TaskRepository taskRepository;
+
     private final PasswordEncoder passwordEncoder;
 
-    public UserService(UserRepository userRepository, PasswordEncoder passwordEncoder) {
+    public UserService(
+            UserRepository userRepository,
+            TaskRepository taskRepository,
+            PasswordEncoder passwordEncoder
+    ) {
         this.userRepository = userRepository;
+        this.taskRepository = taskRepository;
         this.passwordEncoder = passwordEncoder;
     }
 
@@ -63,6 +72,11 @@ public class UserService {
 
     public void delete(Long id) {
         var user = getById(id);
+
+        if (taskRepository.existsByAssignee(user)) {
+            throw new ResourceConflictException("User has tasks");
+        }
+
         userRepository.delete(user);
     }
